+++ lib/slack-ruby-bot/bot.rb
@@ -9,5 +9,9 @@ module SlackRubyBot
     def self.instance
       SlackRubyBot::App.instance
     end
+
+    def self.call(client, data, _match)
+      client.say(channel: data.channel, text: "Sorry <@#{data.user}>, I don't understand that command!", gif: 'idiot')
+    end
   end
 end